requirements.txt
langchain==0.1.16
pymongo==4.6.1
langchain-community==0.0.34
langchain-openai==0.1.3
langgraph
tavily-python
langchain-groq
starlette==0.37.2
uvicorn==0.29.0
psycopg==3.1.18
psycopg-binary==3.1.18
python-ulid==2.5.0
magentic